Major Pre-Season Updates. Implemented a new easy hot-swappable control system for both 1D (single motors) and 2D (Drivetrain) systems. Implemented 1D Control Methods HalfStick and SingleButton Implemented 2D Control Methods FullStick.
Still Planned:
DoubleButton (uses two different buttons and outputs to a 1D control).
HatConstantHalf (uses the DPAD as essentially a stick that outputs only 0, -1, 1 or scaled values. Outputs 1D control)
HatConstantFull (uses the DPAD as essentially a stick that outputs only 0, -1, 1 or scaled values. Outputs 2D control).
HalfStickButtonSwitch (Works similarly to a HalfStick but also has a button that can be used to switch between two different configs.)
FullStickButtonSwitch (Works similarly to a FullStick but also has a button that can be used to switch between two different configs.)