Removed Current Command System to Prepare for Update to WPI NewCommandLibrary Removed Shooter and ShooterRotator as they are no longer used and likely not to be. Updated both DriveTrain classes to no longer be Subsystems until NewCommands is added. Updated Robot to reflect module removal. Updated ControllerManager to support chaining the init method. Updated RobotMap to reflect the changes to modules and ControllerManger (init now takes place here, fixes bug where the controller passed to the default controller method was null) Updated Robot to remove ControllerManager init method there.
Still Planned:
HatConstantHalf (uses the DPAD as essentially a stick that outputs only 0, -1, 1 or scaled values. Outputs 1D control)
HatConstantFull (uses the DPAD as essentially a stick that outputs only 0, -1, 1 or scaled values. Outputs 2D control).
HalfStickButtonSwitch (Works similarly to a HalfStick but also has a button that can be used to switch between two different configs.)
FullStickButtonSwitch (Works similarly to a FullStick but also has a button that can be used to switch between two different configs.)