Integrated the POV/DPad/Hat into StickMode
Implemented a basic Command System with a couple of basic Commands. This will be expanded upon.

Implemented Claw and finished Arm to reflect robot changes.

Implemented CombinedDouble to allow for a simple combination of two SingleControlStyles into a DoubleControlStyle

Added all Subsystems into a new Subsystem class. This is used by the new Scheduler to prevent two commands from attempting to access to the same part at the same time.

Still Planned:

HalfStickButtonSwitch (Works similarly to a HalfStick but also has a button that can be used to switch between two different configs.)
FullStickButtonSwitch (Works similarly to a FullStick but also has a button that can be used to switch between two different configs.)
Push configs from robot environment.